requirements.txt
Django==5.2
python-dotenv
requests
edge-tts
pydub
googletrans==4.0.0rc1
libretranslatepy
yt-dlp
pytube
openai-whisper
pdf2image
pytesseract
librosa
noisereduce
numpy==1.23.5
scipy
audioread
soundfile
matplotlib